"Contact\User" class created
