+++ view/theme/vier/theme.php
@@ -117,7 +117,7 @@ function vier_community_info()
 
 	// comunity_profiles
 	if ($show_profiles) {
-		$contacts = Contact::getSuggestions(local_user(), 0, 9);
+		$contacts = Contact\Relation::getSuggestions(local_user(), 0, 9);
 
 		$tpl = Renderer::getMarkupTemplate('ch_directory_item.tpl');
 		if (DBA::isResult($contacts)) {
